changed latitude and longitude to Float! in type def

+++ server/schemas/typeDefs.js
@@ -17,8 +17,8 @@ const typeDefs = gql`
       title: String
       description: String
       rating: Int
-      latitude: Int
-      longitude: Int
+      latitude: Float
+      longitude: Float
 
    }
